% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuity.R
\name{calcContinuityFromDist}
\alias{calcContinuityFromDist}
\title{Calculate continuity based on distance matrices}
\usage{
calcContinuityFromDist(distOriginal, distLowDim, kTM)
}
\arguments{
\item{distOriginal}{N x N matrix or dist object, representing pairwise sample
distances based on the original (high-dimensional) observed values.}

\item{kTM}{The number of nearest neighbors}

\item{rankLowDim}{N x N matrix or dist object, representing pairwise sample
distances based on the low-dimensional representation.}
}
\value{
The continuity value.
}
\description{
The continuity was proposed by Kaski et al, as a local quality measure
of a low-dimensional representation. The metric focuses on the preservation
of local neighborhoods, and compares the neighborhoods of points in the
low-dimensional representation to those in the original data. Hence, the
continuity measure indicates to which degree we can trust that the points
closest to a given sample in the original data set are placed close to the
sample also in the low-dimensional representation. The \code{kTM} parameter
defines the size of the neighborhoods to consider.
}
\references{
Kaski, S., Nikkilä, J., Oja, M., Venna, J., Törönen, P., and
  Castrén, E. (2003). Trustworthiness and metrics in visualizing similarity
  of gene expression. BMC Bioinformatics 4:48.
}
\author{
Charlotte Soneson
}
\keyword{internal}
