% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuity.R
\name{calcContinuityFromRank}
\alias{calcContinuityFromRank}
\title{Calculate continuity based on sample rankings}
\usage{
calcContinuityFromRank(rankReference, rankLowDim, kTM)
}
\arguments{
\item{rankReference}{N x N matrix, each row/column corresponding to one
sample. The value of entry (i, j) represents the position of sample i in
the ranking of all samples with respect to their distance from sample j,
based on the reference (high-dimensional) observed values. The most similar
sample (i.e., sample j itself) has position 1.}

\item{rankLowDim}{N x N matrix, each row/column corresponding to one sample.
The value of entry (i, j) represents the position of sample i in the
ranking of all samples with respect to their distance from sample j, based
on the low-dimensional representation. The most similar sample (i.e.,
sample j itself) has position 1.}

\item{kTM}{The number of nearest neighbors}
}
\value{
The continuity value.
}
\description{
The continuity was proposed by Venna and Kaski, as a local quality measure of
a low-dimensional representation. The metric focuses on the preservation of
local neighborhoods, and compares the neighborhoods of points in the
low-dimensional representation to those in the reference data. Hence, the
continuity measure indicates to which degree we can trust that the points
closest to a given sample in the reference data set are placed close to the
sample also in the low-dimensional representation. The \code{kTM} parameter
defines the size of the neighborhoods to consider.
}
\references{
Venna J., Kaski S. (2001). Neighborhood preservation in nonlinear
  projection methods: An experimental study. In Dorffner G., Bischof H.,
  Hornik K., editors, Proceedings of ICANN 2001, pp 485–491. Springer,
  Berlin.
}
\author{
Charlotte Soneson
}
\keyword{internal}
