% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trustworthiness.R
\name{calcTrustworthinessFromRank}
\alias{calcTrustworthinessFromRank}
\title{Calculate trustworthiness based on sample rankings}
\usage{
calcTrustworthinessFromRank(rankOriginal, rankLowDim, kTM)
}
\arguments{
\item{rankOriginal}{N x N matrix, each row/column corresponding to one
sample. The value of entry (i, j) represents the position of sample i in
the ranking of all samples with respect to their distance from sample j,
based on the original (high-dimensional) observed values. The most similar
sample (i.e., sample j itself) has position 1.}

\item{rankLowDim}{N x N matrix, each row/column corresponding to one sample.
The value of entry (i, j) represents the position of sample i in the
ranking of all samples with respect to their distance from sample j, based
on the low-dimensional representation. The most similar sample (i.e.,
sample j itself) has position 1.}

\item{kTM}{The number of nearest neighbors}
}
\value{
The trustworthiness value.
}
\description{
The trustworthiness was proposed by Kaski et al, as a local quality measure
of a low-dimensional representation. The metric focuses on the preservation
of local neighborhoods, and compares the neighborhoods of points in the
low-dimensional representation to those in the original data. Hence, the
trustworthiness measure indicates to which degree we can trust that the
points placed closest to a given sample in the low-dimensional representation
are really close to the sample also in the original data set. The \code{kTM}
parameter defines the size of the neighborhoods to consider.
}
\references{
Kaski, S., Nikkilä, J., Oja, M., Venna, J., Törönen, P., and
  Castrén, E. (2003). Trustworthiness and metrics in visualizing similarity
  of gene expression. BMC Bioinformatics 4:48.
}
\author{
Charlotte Soneson
}
\keyword{internal}
