% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreval-pkg.R
\docType{package}
\name{dreval-package}
\alias{dreval-package}
\title{dreval}
\description{
dreval evaluates and compares multiple reduced dimension
representations, based on how well they retain structure from
the original data set.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/csoneson/dreval}
  \item Report bugs at \url{https://github.com/csoneson/dreval/issues}
}

}
\author{
\strong{Maintainer}: Charlotte Soneson \email{charlottesoneson@gmail.com} (\href{https://orcid.org/0000-0003-3833-2169}{ORCID})

}
\keyword{internal}
