% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreval-pkg.R
\docType{package}
\name{dreval-pkg}
\alias{dreval-pkg}
\title{dreval}
\description{
dreval evaluates and compares multiple reduced dimension
representations, based on how well they retain structure from
the original data set.
}
