% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmc3ksub.R
\docType{data}
\name{pbmc3ksub}
\alias{pbmc3ksub}
\title{Expression profiles for 2,700 PBMCs}
\description{
This data set contains expression profiles for 2,700 PBMCs. The original data
set was obtained from the TENxPBMCData package (pbmc3k data set). This data
set was subset to around 1,800 highly expressed genes, normalized using the
scran package, and several dimensionality reduction methods were applied.
}
\author{
Charlotte Soneson
}
\keyword{data}
