% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dreval.R
\name{dreval}
\alias{dreval}
\title{Evaluate dimension reduction methods}
\usage{
dreval(sce, dimReds = NULL, assay = "logcounts", features = NULL,
  nSamples = NULL, distNorm = FALSE, kTM = c(10, 100),
  verbose = FALSE)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{dimReds}{A character vector giving the names of the reduced dimension
representations from \code{sce} to compare.}

\item{assay}{A character scalar giving the name of the assay from the
\code{sce} to use as the basis for the comparisons.}

\item{features}{A character vector giving the IDs of the features to use for
distance calculations from the chosen assay.}

\item{nSamples}{A numeric scalar, the number of columns to subsample
(randomly) from the \code{sce}.}

\item{distNorm}{A logical scalar, whether the distance vectors should be L2
normalized before the Sammon stress and the Euclidean distance between them
is calculated.}

\item{kTM}{An integer vector giving the number of neighbors to use for
trustworthiness and continuity calculations.}

\item{verbose}{A logical scalar, whether to print out progress messages.}
}
\value{
A \code{data.frame} with values for all evaluation metrics, across
  the dimension reduction methods. The following metrics are included:
  \itemize{
  \item SpearmanCorrEuclDist - The Spearman correlation between the original,
  high-dimensional Euclidean distances and the Euclidean distances in the
  low-dimensional representation.
  \item PearsonCorrEuclDist - The Pearson correlation between the original,
  high-dimensional Euclidean distances and the Euclidean distances in the
  low-dimensional representation.
  \item EuclDistToEuclDist - The Euclidean distance between the vector of
  Euclidean distances in the original, high-dimensional data and those in the
  low-dimensional representation. If \code{distNorm} is TRUE, the Euclidean
  distance vectors are normalized to have an L2 norm equal to 1 before the
  distance between them is calculated. If \code{distNorm} is FALSE, the
  distance vectors are instead divided by the square root of their length, to
  make the value independent of the number of samples.
  \item SammonStress - The Sammon stress. If \code{distNorm} is TRUE, the
  Euclidean distance vectors are normalized to have an L2 norm equal to 1
  before the stress is calculated. If \code{distNorm} is FALSE, the distance
  vectors are instead divided by the square root of their length, to make the
  value independent of the number of samples.
  \item TrustworthinessEuclDist_kNN - The trustworthiness score (Kaski et al.
  2003), using NN nearest neighbors. The trustworthiness indicates to which
  degree we can trust that the points placed closest to a given sample in the
  low-dimensional representation are really close to the sample also in the
  original data set.
  \item ContinuityEuclDist_kNN - The continuity score (Kaski et al. 2003),
  using NN nearest neighbors. The continuity indicates to which degree we can
  trust that the points closest to a given sample in the original data set
  are placed close to the sample also in the low-dimensional representation.
  }
}
\description{
Calculate several goodness-of-fit metrics for one or multiple reduced
dimension representations.
}
\references{
Kaski, S., Nikkilä, J., Oja, M., Venna, J., Törönen, P., and
  Castrén, E. (2003). Trust- worthiness and metrics in visualizing similarity
  of gene expression. BMC Bioinformatics 4:48.
}
\author{
Charlotte Soneson
}
