% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotRankSummary}
\alias{plotRankSummary}
\title{Plot summary of the dimension reduction method ranking across metrics}
\usage{
plotRankSummary(dreSummary, metrics = NULL)
}
\arguments{
\item{dreSummary}{A \code{data.frame} with the values of the evaluation
metrics, typically the \code{"scores"} element of the output of
\code{dreval()}.}

\item{metrics}{A character vector with the metrics to include in the summary.
Must be a subset of the column names of \code{dreSummary}. If \code{NULL},
all metrics will be used.}
}
\value{
Nothing is returned, but a plot is generated.
}
\description{
For each metric, rank the evaluated reduced dimension representations by
performance, and plot a summary of the overall ranking. Metrics evaluating
local and global structure preservations are colored in red and blue,
respectively.
}
\author{
Charlotte Soneson
}
